Move specs to specs folder, configure dynamic msw stubs
