+++ cypress.json
@@ -1,3 +1,9 @@
 {
-  "baseUrl": "http://localhost:3000"
+  "baseUrl": "http://localhost:3000",
+  "fixturesFolder": "specs/fixtures",
+  "integrationFolder": "specs/integration",
+  "pluginsFile": "specs/plugins/index.js",
+  "screenshotsFolder": "specs/screenshots",
+  "videosFolder": "specs/videos",
+  "supportFile": "specs/support/index.js"
 }
