+++ specs/integration/bridgeapi/users/sign_up.spec.js
