+++ specs/integration/examples/actions.spec.js
