+++ specs/integration/examples/aliasing.spec.js
