+++ specs/integration/examples/connectors.spec.js
