+++ specs/integration/examples/cookies.spec.js
