+++ specs/integration/examples/local_storage.spec.js
