+++ specs/integration/examples/location.spec.js
