+++ specs/integration/examples/spies_stubs_clocks.spec.js
