+++ specs/integration/examples/traversal.spec.js
