+++ specs/integration/examples/window.spec.js
