+++ specs/support/mocks/browser.js
